+++ examples/sedk-mysql-js-example/index.js
@@ -1,17 +1,28 @@
 const examples = [
-	'./readme-example.js',
-	'./example1.js',
+    './readme-example.js',
+    './example1.js',
+    './sedk-mysql_v0_0_4.js',
 ]
 
 async function main() {
-	const functions = examples.map((file) => async () => await import(file))
-	for (const fn of functions) {
-		console.log(`📦 run ${fn.name}`)
-		await fn()
-	}
+    const functions = examples.map((filename) => ({
+        filename,
+        execFile: async () => await import(filename)
+    }))
+    for (const fn of functions) {
+        console.log(`📦 run ${fn.filename}`)
+        await fn.execFile()
+    }
 }
 
 console.log('🚀 Starting sedk-mysql-js-example ...')
-main().then(() => {
-	console.log('👋 End of sedk-mysql-js-example')
-})
+main()
+    .then(() => {
+        console.log()
+        console.log('👋 End of sedk-mysql-js-example')
+    })
+    .catch((err) => {
+        console.log('❌ Error')
+        console.error(err)
+        process.exit(1)
+    })