+++ packages/sedk-mysql/src/errors.ts
@@ -21,6 +21,18 @@ export class DeleteWithoutConditionError extends Error {
 	}
 }
 
+export class InvalidLimitValueError extends Error {
+	constructor(value: number) {
+		super(`Invalid limit value: ${value}, value must be positive integer number`)
+	}
+}
+
+export class InvalidOffsetValueError extends Error {
+	constructor(value: number) {
+		super(`Invalid offset value: ${value}, value must be positive integer number`)
+	}
+}
+
 export class InsertColumnsAndValuesNotEqualError extends Error {
 	constructor(columnsCount: number, valuesCount: number) {
 		super()