+++ packages/sedk-mysql/test/exported/branches.test.ts
@@ -5,6 +5,7 @@ import { database } from '@test/database'
 const table1 = database.s.public.t.table1
 const col1 = table1.c.col1
 const col3 = table1.c.col3
+const table2 = database.s.public.t.table2
 
 describe('Branches', () => {
 	const sql = builder(database)
@@ -18,12 +19,12 @@ describe('Branches', () => {
 			expect(whereStep1.getSQL()).toEqual("SELECT `col1` FROM `table1` WHERE `col1` = 'x1';")
 			expect(whereStep2.getSQL()).toEqual("SELECT `col1` FROM `table1` WHERE `col1` = 'x2';")
 		})
-		it.skip('produce 2 different statements for 2 WHERE steps added after ON step', () => {
+		it('produce 2 different statements for 2 WHERE steps added after ON step', () => {
 			const fromStep = sql
 				.select(col1)
 				.from(table1)
-				// .leftJoin(table2)
-				// .on(col1.eq(table2.c.col1))
+				.leftJoin(table2)
+				.on(col1.eq(table2.c.col1))
 
 			const whereStep1 = fromStep.where(col3.eq('x1'))
 			const whereStep2 = fromStep.where(col3.eq('x2'))