+++ packages/sedk-mysql/test/unit/steps/select-path/AfterFromStep.test.ts
@@ -1,8 +1,0 @@
-import { RootStep } from '@src/steps'
-import { builderData } from '@test/builderData'
-import { database } from '@test/database'
-
-//Aliases
-const table1 = database.s.public.t.table1
-
-it.todo('AfterFromStep')