+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.3.2]  2023-11-19
+### Changed
+- Lint must have no warning in GitHub pipeline
+- Update root package dependencies to latest version (sedk only has dev dependencies), which fix vulnerabilities
+- Remove unused dependencies `ts-node`
+- End support for node v14.x (only applicable for root package)
+- publish sedk-mysql in pipeline
+
 ## [0.3.1]  2023-09-23
 ### Added
 - test against node v20.x in pipeline