+++ packages/sedk-mysql/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.0.4]  2023-11-16
+###
+- support join, leftJoin, rightJoin and innerJoin steps.
+- support limit and offset steps.
+
 ## [0.0.3]  2023-10-04
 ### Fix
 - earlier version publish